Hide fbclid FB's tracking param if present
I don't need to redirect to completely remove the param, hiding it is enough. Mostly done with JS because I like the URL API and removing just the fbclid param in plain PHP is a bit difficult. I miss my framework not used here.

Here's how it works https://www.michalspacek.com/using-javascript-to-modify-urls-and-hide-fbclid