Use Standards mode
The pages were initially designed to be as simple as possible HTML-wise so people can view source code. But since then relevant JS is shown on-click in the page not only for mobile devices, so understanding the HTML source is not a relevant problem to solve.

Also, Firefox complains in devtools console that «This page is in Quirks Mode. Page layout may be impacted. For Standards Mode use “<!DOCTYPE html>”»