Switch to using and explaining Reporting-Endpoints header, not Report-To
Except for NEL as it doesn't support Reporting API v1 (Reporting-Endpoints).

Report-To is deprecated and Chrome supports Reporting-Endpoints for a few years already. Well, except for NEL, it doesn't send all NEL reports with Reporting-Endpoints. It does send some (like tcp.refused when the site goes down), but not DNS errors and 404s.

Close #21